% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigancolors.R
\name{michigancolors}
\alias{michigancolors}
\title{University of Michigan color palette}
\usage{
michigancolors(x = "all")
}
\arguments{
\item{x}{colors. Defaults to all. Allowable colors are: "blue", "maize", "tappan red", "ross school orange", "rackham green", "wave field green", "taubman teal", "arboretum blue", "ann arbor amethyst", "matthaei violet", "umma tan", "burton tower beige", "angell hall ash", and "law quad stone".}
}
\description{
Use University of Michigan colors in graphs.
}
\examples{
hist(iris$Petal.Width, col = michigancolors())
hist(iris$Petal.Length, col = michigancolors("blue"))
plot(iris$Petal.Length, iris$Petal.Width, col = michigancolors("tappan red"), pch = 19)
}
\keyword{Michigan}
\keyword{colors}
