% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigancolors.R
\name{michigancolors}
\alias{michigancolors}
\title{University of Michigan color palette}
\usage{
michigancolors(x = "all")
}
\arguments{
\item{x}{colors. Defaults to all.}
}
\description{
Use University of Michigan colors in graphs.
}
\examples{
hist(iris$Petal.Length, col = michigancolors())
hist(iris$Petal.Width, col = michigancolors())
plot(iris$Petal.Length, iris$Petal.Width, col = michigancolors("blue"))
}
\keyword{Michigan}
\keyword{colors}
