% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigancolors.R
\name{michigancolors}
\alias{michigancolors}
\title{University of Michigan color palette}
\usage{
michigancolors(x = "all")
}
\arguments{
\item{x}{colors. Defaults to all.}
}
\description{
Use University of Michigan colors in graphs.
}
\examples{
hist(iris$Petal.Length, col = michigancolors())
hist(x, col = michigancolors())
plot(x, y, col = michigancolors("blue"))
Allowable colors are: "blue", "maize", "tappan red",
"ross school orange", "wave field green",
"taubman teal", "arboretum blue", and
"ann arbor amethyst"
}
\keyword{Michigan}
\keyword{colors}
